requirements.txt
streamlit==1.40.1
scikit-learn==1.6.1
pandas==2.0.3
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.0.1
